force ADVICE_DIR to default to our local test advices

+++ tests/advice/shouty/__init__.py
@@ -7,7 +7,9 @@ class Check(BaseCheck):
     order = 10
     def run(self):
         p = (self.env.path / "README.txt")
-        cur = p.read_text()
-        new = cur.upper()
-        p.write_text(new)
-        return cur != new
+        if p.is_file():
+            cur = p.read_text()
+            new = cur.upper()
+            p.write_text(new)
+            return cur != new
+        return False
